[WEB-04] Adding prototype arrays
